add: 診断演算機能の追加
