+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/kitadaikai/Desktop/develop/hackz/hackz_plesio_front/src/components/ExecuteButton.jsx":"1","/Users/kitadaikai/Desktop/develop/hackz/hackz_plesio_front/src/App.js":"2","/Users/kitadaikai/Desktop/develop/hackz/hackz_plesio_front/src/components/QuestionView.jsx":"3"},{"size":177,"mtime":1608395571911,"results":"4","hashOfConfig":"5"},{"size":791,"mtime":1608394843718,"results":"6","hashOfConfig":"5"},{"size":2604,"mtime":1608391161741,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1n6y1w8",{"filePath":"10","messages":"11","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":"12","usedDeprecatedRules":"13"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":3,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/kitadaikai/Desktop/develop/hackz/hackz_plesio_front/src/components/ExecuteButton.jsx",[],"/Users/kitadaikai/Desktop/develop/hackz/hackz_plesio_front/src/App.js",["16","17"],"import React, { useState } from 'react';\nimport './App.css';\nimport QuestionView from './components/QuestionView'\nimport ExecuteButton from './components/ExecuteButton';\n\n// let phase = 1;\n\nconst yes = () => {\n  console.log(\"yes\");\n}\nconst no = () => {\n  console.log(\"no\");\n}\n\n\nfunction App() {\n\n  const [phase, setPhase] = useState(1)\n  const [answers, setAnswers] = useState([null, null, null, null, null, null, null, null, null, null]) // 回答結果を保存するstate\n\n\n  const test = () => {\n    setPhase(phase + 1);\n    console.log(phase)\n  }\n\n  return (\n    <div>\n      { phase === 1 \n        ? <ExecuteButton func={test} title='診断開始' />\n        : null\n      }\n      \n\n      <QuestionView setAnswers={setAnswers} answers={answers}/>\n    </div>\n  );\n}\n\n\nexport default App;\n",["18","19"],"/Users/kitadaikai/Desktop/develop/hackz/hackz_plesio_front/src/components/QuestionView.jsx",["20","21","22"],{"ruleId":"23","severity":1,"message":"24","line":8,"column":7,"nodeType":"25","messageId":"26","endLine":8,"endColumn":10},{"ruleId":"23","severity":1,"message":"27","line":11,"column":7,"nodeType":"25","messageId":"26","endLine":11,"endColumn":9},{"ruleId":"28","replacedBy":"29"},{"ruleId":"30","replacedBy":"31"},{"ruleId":"23","severity":1,"message":"32","line":1,"column":17,"nodeType":"25","messageId":"26","endLine":1,"endColumn":25},{"ruleId":"23","severity":1,"message":"33","line":3,"column":8,"nodeType":"25","messageId":"26","endLine":3,"endColumn":21},{"ruleId":"23","severity":1,"message":"34","line":20,"column":11,"nodeType":"25","messageId":"26","endLine":20,"endColumn":22},"no-unused-vars","'yes' is assigned a value but never used.","Identifier","unusedVar","'no' is assigned a value but never used.","no-native-reassign",["35"],"no-negated-in-lhs",["36"],"'useState' is defined but never used.","'ExecuteButton' is defined but never used.","'Calculation' is assigned a value but never used.","no-global-assign","no-unsafe-negation"]
+[{"/Users/hiratakaiki/develop/hackz/hackz_plesio_front/src/App.js":"1","/Users/hiratakaiki/develop/hackz/hackz_plesio_front/src/components/QuestionView.jsx":"2","/Users/hiratakaiki/develop/hackz/hackz_plesio_front/src/reportWebVitals.js":"3","/Users/hiratakaiki/develop/hackz/hackz_plesio_front/src/components/ExecuteButton.jsx":"4","/Users/hiratakaiki/develop/hackz/hackz_plesio_front/src/components/Calculate.jsx":"5","/Users/hiratakaiki/develop/hackz/hackz_plesio_front/src/index.js":"6"},{"size":942,"mtime":1608427489933,"results":"7","hashOfConfig":"8"},{"size":1062,"mtime":1608427722109,"results":"9","hashOfConfig":"8"},{"size":362,"mtime":1608348628564,"results":"10","hashOfConfig":"8"},{"size":177,"mtime":1608404322978,"results":"11","hashOfConfig":"8"},{"size":1453,"mtime":1608427939554,"results":"12","hashOfConfig":"8"},{"size":500,"mtime":1608348628563,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"11hza2w",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"27"},"/Users/hiratakaiki/develop/hackz/hackz_plesio_front/src/App.js",[],"/Users/hiratakaiki/develop/hackz/hackz_plesio_front/src/components/QuestionView.jsx",[],"/Users/hiratakaiki/develop/hackz/hackz_plesio_front/src/reportWebVitals.js",[],["28","29"],"/Users/hiratakaiki/develop/hackz/hackz_plesio_front/src/components/ExecuteButton.jsx",[],"/Users/hiratakaiki/develop/hackz/hackz_plesio_front/src/components/Calculate.jsx",[],"/Users/hiratakaiki/develop/hackz/hackz_plesio_front/src/index.js",[],["30","31"],{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","replacedBy":"35"},{"ruleId":"32","replacedBy":"36"},{"ruleId":"34","replacedBy":"37"},"no-native-reassign",["38"],"no-negated-in-lhs",["39"],["38"],["39"],"no-global-assign","no-unsafe-negation"]
