+++ src/components/Questions.json
@@ -18,6 +18,8 @@
             "title": "趣味がありますか？"
         }, {
             "title": "よく映画を見ますか？"
+        }, {
+            "title": "音楽は好きですか？"
         }
     ]
 }
